Update README.md
I designed a water-tight lid for the ABS pipe in SolidWorks that uses 4 latches. There are 4 holes made slightly smaller than the cables and tubes going into the pipe on top of the lid: One inlet, one outlet, one pump wire, and one heater wire. 

In addition, a 2mm silicone gasket is used to make a seal between the pipe. Fine sanding on the 3D printed lid on the gasket side (could make a better water seal) and a rough sanding on the sections where the tubes/cables will be glued (to make the glue stick more) would be better.

I first glue the cable and tube with cotton + superglue mix to make it sturdy, and then fill around it with hot glue. This time, however, I used silicone for the inner side. Using silicone for this is not recommended, and I regret using it. The silicone is way too soft and doesn't do its job after the tube/cable bends a little bit.

Filter material: coarse sponge -> fine sponge -> porous stone/ball (for beneficial bacteria)

Recommendation: Use a 3" pipe. A 2" pipe is hard to work with in terms of putting the filter media in.

Material link:
-Silicone gasket: https://www.aliexpress.com/item/1005006259312869.html?spm=a2g0o.order_list.order_list_main.5.40801802B0RrWW

I usually find cheaper items in bundle deals on Aliexpress. Bundle deals don't have links, so I can only leave the silicone gasket, as I didn't buy that one from bundle deals.